SOCINT-311 Refactor Utils Further changes to group items together
