+++ app/register_handlers.py
@@ -12,9 +12,7 @@ from .security import invalidate
 from .session import get_existing_session, get_session_value
 from .utils import View, FlashMessage
 from .service_calls.rhsvc import RHSvcRegisterCase
-from .validators.date_of_birth import ProcessDOB
-from .validators.mobile import ProcessMobileNumber
-from .validators.name import ProcessName
+from .validators.identity import IdentityValidators
 from .exceptions import InvalidDataError, InvalidDataErrorWelsh, TooManyRequestsRegister
 
 logger = get_logger('respondent-home')
@@ -166,7 +164,7 @@ class RegisterEnterName(View):
 
         data = await request.post()
 
-        form_valid = ProcessName.validate_name(request, data, display_region)
+        form_valid = IdentityValidators.validate_name(request, data, display_region)
 
         if not form_valid:
             logger.info('form submission error',
@@ -231,7 +229,7 @@ class RegisterEnterMobile(View):
         data = await request.post()
 
         try:
-            mobile_number = ProcessMobileNumber.validate_uk_mobile_phone_number(data['request-mobile-number'],
+            mobile_number = IdentityValidators.validate_uk_mobile_phone_number(data['request-mobile-number'],
                                                                                 locale)
 
             logger.info('valid mobile number',
@@ -447,7 +445,7 @@ class RegisterEnterChildName(View):
 
         data = await request.post()
 
-        form_valid = ProcessName.validate_name(request, data, display_region, child=True)
+        form_valid = IdentityValidators.validate_name(request, data, display_region, child=True)
 
         if not form_valid:
             logger.info('form submission error',
@@ -592,7 +590,7 @@ class RegisterChildDOB(View):
         data = await request.post()
 
         try:
-            date = ProcessDOB.validate_dob(data)
+            date = IdentityValidators.validate_dob(data)
 
             logger.info('valid dob',
                         client_ip=request['client_ip'], client_id=request['client_id'], trace=request['trace'])
@@ -646,8 +644,8 @@ class RegisterChildSummary(View):
                                               'child_middle_names', user_journey, request_type),
             'last_name': get_session_value(request, register_attributes, 'child_last_name', user_journey, request_type),
             'school_name': get_session_value(request, register_attributes, 'school_name', user_journey, request_type),
-            'child_dob': ProcessDOB.format_dob(get_session_value(request, register_attributes, 'child_dob',
-                                                                 user_journey, request_type))
+            'child_dob': IdentityValidators.format_dob(get_session_value(request, register_attributes, 'child_dob',
+                                                                         user_journey, request_type))
         }
 
     async def post(self, request):
