+++ app/validators/address.py
@@ -6,7 +6,7 @@ from app.exceptions import InvalidDataError, InvalidDataErrorWelsh
 from app import validators
 
 
-class ProcessPostcode:
+class AddressValidators:
     postcode_validation_pattern = re.compile(
         r'^((AB|AL|B|BA|BB|BD|BH|BL|BN|BR|BS|BT|BX|CA|CB|CF|CH|CM|CO|CR|CT|CV|CW|DA|DD|DE|DG|DH|DL|DN|DT|DY|E|EC|EH|EN|EX|FK|FY|G|GL|GY|GU|HA|HD|HG|HP|HR|HS|HU|HX|IG|IM|IP|IV|JE|KA|KT|KW|KY|L|LA|LD|LE|LL|LN|LS|LU|M|ME|MK|ML|N|NE|NG|NN|NP|NR|NW|OL|OX|PA|PE|PH|PL|PO|PR|RG|RH|RM|S|SA|SE|SG|SK|SL|SM|SN|SO|SP|SR|SS|ST|SW|SY|TA|TD|TF|TN|TQ|TR|TS|TW|UB|W|WA|WC|WD|WF|WN|WR|WS|WV|YO|ZE)(\d[\dA-Z]?[ ]?\d[ABD-HJLN-UW-Z]{2}))$'  # NOQA
     )
@@ -44,7 +44,7 @@ class ProcessPostcode:
             else:
                 raise InvalidDataError('Enter a valid UK postcode')
 
-        if not ProcessPostcode.postcode_validation_pattern.fullmatch(postcode):
+        if not AddressValidators.postcode_validation_pattern.fullmatch(postcode):
             if locale == 'cy':
                 raise InvalidDataErrorWelsh("Rhowch god post dilys yn y Deyrnas Unedig")
             else:
